requirements.txt
absl-py==0.15.0
aiohttp==3.8.6
aiosignal==1.2.0
alembic==1.7.7
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autopage==0.5.2
banal==1.0.6
beautifulsoup4==4.12.2
blis==0.7.11
Bottleneck==1.3.7
bs4==0.0.2
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==2.0.12
ci-info==0.2.0
clang==5.0
click==8.0.4
cliff==3.10.1
cmaes==0.9.0
cmd2==2.4.3
colorlog==6.8.0
confection==0.1.3
configobj==5.0.8
configparser==5.2.0
contextvars==2.4
cycler==0.11.0
cymem==2.0.8
Cython==3.0.5
dataclasses==0.8
dataset==1.6.2
datasets==2.4.0
decorator==4.4.2
dill==0.3.4
etelemetry==0.2.2
fastai==2.5.5
fastcore==1.4.0
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.4.1
fitz==0.0.1.dev2
Flask==2.0.3
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.48.2
h5py==3.1.0
httplib2==0.22.0
huggingface-hub==0.4.0
idna==3.4
idna-ssl==1.1.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
jsonify==0.5
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
lxml==5.1.0
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.10
networkx==2.5.1
nibabel==3.2.2
nipype==1.7.1
numexpr==2.8.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
openai==0.8.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
optuna==3.0.6
packaging==21.3
pandas==1.1.5
pathlib==1.0.1
pathy==0.10.3
pbr==6.0.0
pdfminer==20191125
Pillow==8.4.0
preshed==3.0.9
prettytable==2.5.0
protobuf==3.19.6
prov==2.0.0
pyarrow==6.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pydantic==1.9.2
pydot==1.4.2
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==5.0.0
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.17.0
rsa==4.9
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
simplejson==3.19.2
six==1.15.0
smart-open==6.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.51
srsly==2.4.8
stevedore==3.5.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
thinc==8.1.12
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
traits==6.4.1
transformers==4.18.0
typer==0.9.0
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.18
wasabi==1.1.2
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.12.1
xxhash==3.2.0
yarl==1.7.2
zipp==3.6.0
